Implementando ControladorFicha

+++ src/br/com/ru/exceptions/SaldoInsuficienteException.java
@@ -3,18 +3,18 @@ package br.com.ru.exceptions;
 public class SaldoInsuficienteException extends Exception{
     
 	private static final long serialVersionUID = 1L;
-	public int saldo;
+	public double saldo;
 
-    public SaldoInsuficienteException(int saldo) {
+    public SaldoInsuficienteException(double saldo) {
         super("O Saldo é insuficiente para a compra!");
         this.saldo = saldo;
     }
 
-    public int getSaldo() {
+    public double getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(int saldo) {
+    public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
 }
